Restructuring compiling routine
Previously foil module compilation was coupled to the loader API, now it's separated and made its own routine for the sake of setting up a watch mode.

There's still work to be done on the reference backends/frontends, which will most likely use React hooks, Rust, Postgres, make merging the backend with the daemon trivial.